+++ duckscript_sdk/src/sdk/std/fs/dirname/mod.rs
@@ -1,0 +1,39 @@
+use crate::utils::{io, pckg};
+use duckscript::types::command::{Command, CommandResult};
+
+#[cfg(test)]
+#[path = "./mod_test.rs"]
+mod mod_test;
+
+struct CommandImpl {
+    package: String,
+}
+
+impl Command for CommandImpl {
+    fn name(&self) -> String {
+        pckg::concat(&self.package, "GetParentDirectory")
+    }
+
+    fn aliases(&self) -> Vec<String> {
+        vec!["dirname".to_string()]
+    }
+
+    fn help(&self) -> String {
+        include_str!("help.md").to_string()
+    }
+
+    fn run(&self, arguments: Vec<String>) -> CommandResult {
+        if arguments.is_empty() {
+            CommandResult::Error("Path not provided.".to_string())
+        } else {
+            let parent_path = io::get_parent_directory_name(&arguments[0]);
+            CommandResult::Continue(parent_path)
+        }
+    }
+}
+
+pub(crate) fn create(package: &str) -> Box<dyn Command> {
+    Box::new(CommandImpl {
+        package: package.to_string(),
+    })
+}