+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### v0.1.2
 
+* New **dirname** command #6
 * New **canonicalize** command #21
 * New **basename** command #5
 * New **mkdir** command #13